Stash 'dist_auth', a normalized "from:auth_id" string
This cleans up the display of from:zef modules, which previously are
displayed as "Nifty::Module zef:zef:clever-user".

Sensibly, the META6.json "auth" field often (should?) contains the
source prefix in it, like "auth": "zef:some-user". But not always; the
CPAN modules have "auth" injected as just the PAUSE id, for example.
So the best we can do here is strip the source prefix if it is an
exact match.