[REAPP] Update IRC links from freenode to libera
This doesn't change numerous instances of Zoffix on freenode in the contact
information sections of the POD. It specifically says "original version ...
written by", which is still accurate. I don't know if he's on libera, or IRC
at all, and didn't know what to do with this, so just left it.